ctas7-voice-enterprise/pyproject.toml
[project]
name = "ctas7-voice-enterprise"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Charlie Payne", email = "usneodcp@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "elevenlabs>=2.16.0",
    "fastapi>=0.118.0",
    "prometheus-client>=0.23.1",
    "pydantic>=2.12.0",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "structlog>=25.4.0",
    "uvicorn>=0.37.0",
    "websockets>=15.0.1",
]

[project.scripts]
ctas7-voice-enterprise = "ctas7_voice_enterprise:main"

[build-system]
requires = ["uv_build>=0.8.3,<0.9.0"]
build-backend = "uv_build"
